# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
      "AdminAlertDto" : {
        "description" : "One row of the admin alerts (newest first). The ring is in-memory — it clears on a backend restart, so this is a triage view, not a durable log.",
        "properties" : {
          "at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kind" : {
            "description" : "The closed vocabulary of ThrottleAlert: submission daily cap / OTP contact cap / near-duplicate.",
            "type" : "string"
          },
          "retryAfterSeconds" : {
            "description" : "Present only for the 429 alerts (the seconds the caller must wait).",
            "format" : "int32",
            "type" : "integer"
          },
          "subject" : {
            "description" : "The flagged account or contact (user:<id> / contact:<normalized>).",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "AdminAuditDto" : {
        "description" : "One row of the moderation audit trail — newest first. Every moderation-relevant action (admin AND automatic AUTO_CONFIRM).",
        "properties" : {
          "action" : {
            "enum" : [ "STATUS_CHANGE", "DELETE", "REPORT_DISMISS", "CONFIRM", "AUTO_CONFIRM", "REJECT", "USER_SUSPEND", "USER_UNSUSPEND", "MARK_INACCURATE", "CLEAR_INACCURATE", "GUIDANCE_PUBLISH", "GUIDANCE_UNPUBLISH", "GUIDANCE_DELETE", "MEDIA_DELETE" ],
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "moderatorName" : {
            "description" : "The acting admin's profile name; 'Unknown' if the account no longer exists.",
            "type" : "string"
          },
          "newStatus" : {
            "description" : "The review_status value after the action; null for DELETE.",
            "enum" : [ "NEW", "CONFIRMED", "REJECTED" ],
            "type" : "string"
          },
          "previousStatus" : {
            "description" : "The review_status value before the action; equal to newStatus when the action does not move the review state.",
            "enum" : [ "NEW", "CONFIRMED", "REJECTED" ],
            "type" : "string"
          },
          "reason" : {
            "description" : "The REJECT/NEEDS_INFO note when given; null otherwise.",
            "type" : "string"
          },
          "shelterId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "shelterName" : {
            "description" : "Resolved at read time; renders 'Deleted shelter' once the row is gone (the audit row outlives the shelter).",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ErrorResponse" : {
        "properties" : {
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "The `Authorization: Bearer <accessToken>` header read by the JWT filter. A valid token of a SUSPENDED account authenticates nothing (fresh per-request DB lookup — the DB is the truth, never a token claim).",
        "name" : "bearerAuth",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "aleks.bratsun@reaktiiv.com",
      "name" : "Aleks Bratsun (project owner)"
    },
    "description" : "The machine-readable contract of the OpenShelter backend. Companion to the human contracts: context-and-tasks/05-shelter-api.puml (the read contract), the README 'API' section and docs/whitepaper.md. Served at /v3/api-docs in dev/test only (SPRINGDOC_ENABLED); the committed snapshot is docs/api/openapi.json. Every non-2xx response carries the uniform ErrorResponse body.",
    "license" : {
      "name" : "MIT"
    },
    "title" : "OpenShelter API",
    "version" : "0.0.1-SNAPSHOT"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/admin/alerts" : {
      "get" : {
        "description" : "Newest first: the daily submission cap (429), the per-contact OTP cap (429) and the near-duplicate rejection (409). The ring is IN-MEMORY — it clears on a backend restart, so this is a triage view, not a durable log. limit is 1..200, default 50 (anything else 400 — same idiom as /admin/audit).",
        "operationId" : "listAlerts",
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "description" : "Rows to return, 1..200 (default 50; anything else 400).",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AdminAlertDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "The alert rows (newest first)"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AdminAlertDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "limit outside 1..200"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing/invalid/expired bearer token, or a suspended account whose tokens authenticate nothing."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Authenticated but not allowed (admin kind, verified account or authorship)."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unknown resource id — the uniform error body."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Conflict (duplicate or forbidden state) — the uniform error body."
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit / report throttle exceeded — Retry-After in seconds."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "summary" : "The M3 throttle-abuse alerts",
        "tags" : [ "Admin moderation" ],
        "x-admin-only" : {
          "value" : "true"
        }
      }
    },
    "/admin/audit" : {
      "get" : {
        "description" : "Newest first: every moderation-relevant action (admin AND automatic AUTO_CONFIRM) with the shelter name resolved at read time (\"Deleted shelter\" once the row is gone). limit is 1..200, default 100 (anything else 400).",
        "operationId" : "listAudit",
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "description" : "Rows to return, 1..200 (default 100; anything else 400).",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AdminAuditDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "The audit rows (newest first)"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AdminAuditDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "limit outside 1..200"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing/invalid/expired bearer token, or a suspended account whose tokens authenticate nothing."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Authenticated but not allowed (admin kind, verified account or authorship)."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unknown resource id — the uniform error body."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Conflict (duplicate or forbidden state) — the uniform error body."
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit / report throttle exceeded — Retry-After in seconds."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "summary" : "The moderation audit trail",
        "tags" : [ "Admin moderation" ],
        "x-admin-only" : {
          "value" : "true"
        }
      }
    },
...
    "/admin/reports" : {
      "get" : {
        "description" : "Newest first; optional shelterId narrows to one shelter. Carries the reporter's profile name + email — admin-only data, served from /admin/* only.",
        "operationId" : "listShelterReports",
        "parameters" : [ {
          "name" : "shelterId",
          "in" : "query",
          "description" : "Narrow to one shelter (optional).",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AdminShelterReportDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "The report rows (newest first)"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Validation failure — the uniform error body."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing/invalid/expired bearer token, or a suspended account whose tokens authenticate nothing."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Authenticated but not allowed (admin kind, verified account or authorship)."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unknown resource id — the uniform error body."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Conflict (duplicate or forbidden state) — the uniform error body."
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit / report throttle exceeded — Retry-After in seconds."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "summary" : "The shelter report queue",
        "tags" : [ "Admin moderation" ],
        "x-admin-only" : {
          "value" : "true"
        }
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
}